+++ examples/Fullscreen/src/main.cpp
@@ -1,10 +1,10 @@
 #include "emscripten/html5.h"
 #include "entity/entity_manager.hpp"
+#include "fullscreen.hpp"
 #include "game.hpp"
 #include "input_listener.hpp"
 #include "message/message_bus.hpp"
 #include "message/message_payload.hpp"
-#include "simple_game.hpp"
 #include "standard/2d/interpolation/interpolation_system.hpp"
 #include "standard/2d/sprite/sprite_system.hpp"
 #include "standard/2d/webgl2/webgl2_system.hpp"
@@ -20,12 +20,12 @@
 
 int main(int argc, char *argv[]) {
 
-    auto window = JamJar::GetWindow("Sprites");
+    auto window = JamJar::GetWindow("Fullscreen");
     auto context = JamJar::GetCanvasContext();
 
     auto messageBus = new JamJar::MessageBus();
     new JamJar::EntityManager(messageBus);
-    auto game = new SimpleGame(messageBus);
+    auto game = new Fullscreen(messageBus);
     new JamJar::Standard::_2D::WebGL2System(messageBus, window, context);
     new JamJar::Standard::_2D::InterpolationSystem(messageBus);
     new JamJar::Standard::_2D::SpriteSystem(messageBus);